Bug 1075758 - Import a clean copy of ICU source, without doing any post-import fixups/local-patching at all on it.  r=glandium
UltraBlame original commit: 72d5e6070f983351b4662ed263a14ad9af8bea26

+++ intl/icu/source/common/ustring.cpp
@@ -1486,7 +1486,7 @@ u_terminateWChars(wchar_t *dest, int32_t destCapacity, int32_t length, UErrorCod
 
 
 #define STRING_HASH(TYPE, STR, STRLEN, DEREF) \
-    int32_t hash = 0;                         \
+    uint32_t hash = 0;                        \
     const TYPE *p = (const TYPE*) STR;        \
     if (p != NULL) {                          \
         int32_t len = (int32_t)(STRLEN);      \
@@ -1497,7 +1497,7 @@ u_terminateWChars(wchar_t *dest, int32_t destCapacity, int32_t length, UErrorCod
             p += inc;                         \
         }                                     \
     }                                         \
-    return hash
+    return static_cast<int32_t>(hash)
 
 
 U_CAPI int32_t U_EXPORT2
